Retrieves stores from json file and stores in component state
