+++ src/Stores.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 const Stores = () => {
   return (
     <div>
-      <h1>Hello Mexico City!</h1>
+        <h1>Hello Mexico City!</h1>
     </div>
   )
 }
